### 2021-03-14 #3105 @whedon
[PRE REVIEW]: Fear Learning and Anxiety Response (FLARe): Open-source software for remote human fear conditioning experiments.
**Submitting author:** @McGregor14 (<a href="http://orcid.org/0000-0003-0024-7049">Thomas McGregor</a>)
**Repository:** <a href="https://github.com/flare-kcl/flare-app" target ="_blank">https://github.com/flare-kcl/flare-app</a>
**Version:** v1.0.0
