### 2021-03-14 #3105 @arfon
@McGregor14 - this PR should fix your paper https://github.com/flare-kcl/flare-app/pull/110
