### 2021-03-16 #3105 @McGregor14
Hello @oliviaguest 👋

For the app, I can see olade014 has React Native in their list of languages. Though I can't see anyone else, kellerb has listed React and is also comfortable reviewing social sciences. Other people with React in their list of languages include marwahaha and jdalzatec.

For the portal, pgadige is the only person I can see with Django in their list of languages. However, since Django is based on Python, the following people with an interest in psychology may be appropriate: Andytwoods, snastase, JonathanReardon, peircej, and kirstensgithub

Thanks for joining as editor. Let me know if you need any extra info on the project or software.
