+++ paper.bib
@@ -23,7 +23,8 @@
       Number={3},
       Pages={567--580},
       Year={2008},
-      Publisher={Elsevier}
+      Publisher={Elsevier},
+      doi={10.1016/j.actpsy.2007.11.007}
 }
 
 @article{Purves:2019,
@@ -34,7 +35,8 @@
       Volume={123},
       Pages={103475},
       Year={2019},
-      Publisher={Elsevier}
+      Publisher={Elsevier},
+      doi={10.1016/j.brat.2019.103475}
 }
 
 @article{Purves:2020,
@@ -43,5 +45,6 @@
       Author={{Purves}, K. L. and {Krebs}, G. and {McGregor}, T. and {Constantinou}, E. and {Lester}, K. J. and {Barry}, T. J. and {Craske}, M. G. and {Young}, K. S. and {Breen}, G. and {Eley}, T. C.},
       Journal={bioRxiv},
       Year={2020},
-      Publisher={Cold Spring Harbor Laboratory}
+      Publisher={Cold Spring Harbor Laboratory},
+      doi={10.1101/2020.12.18.423504}
 }
