+++ paper.md
@@ -34,10 +34,10 @@ authors:
     affiliation: 1
   - name: Gerome Breen
     orcid: 0000-0003-2053-1792
-    affiliation: “1, 3”
+    affiliation: '1, 3'
   - name: Thalia C. Eley^[Corresponding author]
     orcid: 0000-0001-6458-0700
-    affiliation: “1, 3”
+    affiliation: '1, 3'
 affiliations:
   - name: Social, Genetic and Developmental Psychiatry Centre; Institute of Psychiatry, Psychology & Neuroscience; King’s College London; London, UK.
     index: 1
@@ -49,11 +49,13 @@ date: 15 March 2021
 bibliography: paper.bib
 ---
 
-# Summary & statement of need
+# Statement of need
 
 ‘Fear Learning and Anxiety Response’ (FLARe) is open-source software for conducting remote human fear conditioning studies using visual and auditory stimuli. Fear conditioning is a well-established experimental paradigm that relies on the paired presentation of neutral and aversive stimuli. The FLARe app enables this process through presentations of neutral visual stimuli and aversive auditory stimuli \autoref{fig:Basic fear conditioning trial}. Fear conditioning is a useful paradigm for modelling aspects of the development and exposure-based treatment of anxiety disorders [@Mineka:2008]. However, fear conditioning studies are traditionally carried out in the laboratory, which can limit the speed and quantity of data collection due to: i) limited availability of researcher man-hours, ii) a lack of ability to conduct simultaneous tests on multiple participants, and iii) participant travel requirements. FLARe has been designed to help enable the rapid, large-scale collection of fear conditioning data by removing these barriers associated with in-person testing. Remote fear conditioning using FLARe has been validated against an in-person assessment of fear conditioning [@Purves:2019].
 
 ![Visualisation of the basic task design implemented in fear conditioning phases of the FLARe app. An inter-trial interval precedes each presentation of a visual stimulus. Visual stimuli are sometimes followed by an unpleasant auditory stimulus. Participants must rate how much they expect the auditory stimulus will follow the visual stimulus.\label{fig:Basic fear conditioning trial}](./repo_assets/basic-trial-figure.png){ width=60% }
+
+# Summary
 
 FLARe consists of an Android and iOS compatible smartphone app and associated web portal. The app delivers fear conditioning experiments to participants, while the portal enables researchers to set up experiments and view/export the resulting data. FLARe was designed for use by researchers with and without prior experience conducting fear conditioning studies. The portal \autoref{fig:FLARe portal screenshot} has a Graphical User Interface (GUI) that makes it simple for researchers to set up a range of flexible experiments using a number of purpose-built modules. Each module enables researchers to implement common operations, such as delivering instructions, displaying questionnaire items, or delivering a phase of the fear conditioning task. Researchers simply build their experiment using a combination of these modules and assign participant IDs to the experiment. Once this process is complete, participants can log in using their unique ID, which causes the app to download the settings of the related experiment. The current version of FLARe was developed to make the software open-source, more user friendly, and increase the number of experimental parameters that can be altered by researchers.
 
