### 2021-04-13 #3105 @andytwoods
I am sorry but I do not feel comfortable reviewing this work. React and Docker are core technologies used here, both of which I have no experience with. I just cannot review this article in any sensible amount of time -- it takes days to get familiar with new tech stacks. And please note that there is an interplay between the different technologies that is also burdensome. It' not just the case that I can install the Django app and review that. There's seems to be a React frontend to the Django app that I would have to set up (with which I have no experience), which requires a set of technologies (e.g. live transpiling from ES6) I cannot become familiar with quickly.
With best wishes, Andy.
