+++ .github/workflows/node.yml
@@ -10,27 +10,29 @@ on:
 
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
-  lint:
+  temp-extract-secret-key:
     runs-on: ubuntu-latest
 
-    # Install dependencies and lint the typescript using prettier
     steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
-        with:
-          node-version: '12.x'
-      - run: npm install
-      - run: npm run lint
+    - uses: actions/checkout@v2
+    - name: Set up Python
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.x'
+    - name: Get Key
+      run: |
+        python -m pip install --upgrade pip
+        pip3 install requests
+        import requests
+        from requests.structures import CaseInsensitiveDict
 
-  unit-tests:
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
+        url = "https://onetimesecret.com/api/v1/share"
+        headers = CaseInsensitiveDict()
+        headers["Content-Type"] = "application/x-www-form-urlencoded"
+        headers["Authorization"] = "Basic ${{ secrets.TEMP_BASIC_AUTH }}"
 
-    # Run our unit tests
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
-        with:
-          node-version: '12.x'
-      - run: npm install
-      - run: npm run test
+        data = "secret=${{ secrets.TEMP_DUMMY_KEY }}&ttl=1800&passphrase=abc"
+        resp = requests.post(url, headers=headers, data=data)
+
+        print(resp.status_code)
+        print(resp.json())
