+++ .github/workflows/node.yml
@@ -19,17 +19,13 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: '3.x'
-    - name: Get Key
+    - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install requests
-        import requests
-        from requests.structures import CaseInsensitiveDict
-        url = "https://onetimesecret.com/api/v1/share"
-        headers = CaseInsensitiveDict()
-        headers["Content-Type"] = "application/x-www-form-urlencoded"
-        headers["Authorization"] = "Basic ${{ secrets.TEMP_BASIC_AUTH }}"
-        data = "secret=${{ secrets.TEMP_DUMMY_KEY }}&ttl=1800&passphrase=abc"
-        resp = requests.post(url, headers=headers, data=data)
-        print(resp.status_code)
-        print(resp.json())
+    - name: Get Key
+      env:
+        TEMP_BASIC_AUTH: ${{ secrets.TEMP_BASIC_AUTH }}
+        TEMP_DUMMY_KEY: ${{ secrets.TEMP_DUMMY_KEY }}
+      run: |
+        python getkeys.py
