+++ .github/workflows/node.yml
@@ -24,6 +24,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install requests
     - name: Get Key
+      working-directory: .github/workflows/
       env:
         TEMP_BASIC_AUTH: ${{ secrets.TEMP_BASIC_AUTH }}
         TEMP_DUMMY_KEY: ${{ secrets.TEMP_DUMMY_KEY }}
