### 2021-05-21 #3270 @McGregor14
Hi @oliviaguest,

Sorry for the delay, I am writing up my thesis at the moment so a little bit distracted!

We have now added a license for the app repository (https://github.com/flare-kcl/flare-app/blob/develop/LICENSE) and will add one to the portal repository ASAP.
