### 2021-06-03 #3270 @McGregor14
@oliviaguest portal license is up! https://github.com/flare-kcl/flare-portal/blob/main/LICENSE
