### 2021-06-05 #3270 @oliviaguest
@McGregor14 thanks! Can you clarify if you are happy with explaining to users/reviewers in your documentation how to install this software? I got an email from @EleonoraVagnoni that this is not clear enough at the moment. I personally think this should be accessible (as users to collect data) for people who might not be familiar with the process of installing, e.g., the required background software. What do you think? ☺️
