### 2021-06-14 #3270 @oliviaguest
I think the readme makes more sense, however... I would suggest you discuss this with @EleonoraVagnoni.
