### 2021-06-14 #3270 @McGregor14
Hi @EleonoraVagnoni, Nathan is going to add some instructions to the readme tomorrow. Let us know if there is anything specific you would like him to include or expand upon. I seem to remember the JOSS paper is supposed to stay around a couple of pages at most, so we will keep the instructions to the readme for now but we are happy to move them across if you think it is appropriate.
