### 2021-06-14 #3270 @EleonoraVagnoni
Hi @McGregor14 there isn't anything specific I would like you to include, just some very basic instructions. Sorry about all the trouble. I have tried to install it few times and I was unable to.
