### 2021-06-21 #3270 @NathHorrigan
@EleonoraVagnoni  @olgadoronina  - I have updated the readme with some pointers on how to setup the repo, Let me know if there is anything else I can help with :)
