### 2021-07-01 #3270 @EleonoraVagnoni
> @EleonoraVagnoni @olgadoronina - I have updated the readme with some pointers on how to setup the repo, Let me know if there is anything else I can help with :)

Sorry, can you give me more info regarding how to go to the home directory and create a .bashrc file?
