### 2021-07-01 #3270 @EleonoraVagnoni
I have downloaded Android SDK, I am using a MacOs Big Sur
