### 2021-07-01 #3270 @NathHorrigan
In the terminal if you run:
`touch ~/.bashrc && open ~/.bashrc`

and then paste the code from the Guide that I linked.
