### 2021-07-01 #3270 @EleonoraVagnoni
when I copy and paste react-native init RNEdpresso in the integrated terminal of VS code I receive a message that the command is not found, what am I doing wrong?
