### 2021-07-01 #3270 @NathHorrigan
Ah sorry, You don't need to do that step.
You should be able run `npm i && npm start` in the project repo now.
