### 2021-07-05 #3270 @EleonoraVagnoni
I have tried but it says command not found
