### 2021-07-12 #3270 @NathHorrigan
Hi @EleonoraVagnoni

What about when you run `source ~/.bashrc` in the terminal and then the above commands?
