### 2021-07-12 #3270 @NathHorrigan
If that doesn't work could you try `brew install node` again?
