### 2021-07-14 #3270 @KirstensGitHub
Hi @NathHorrigan I'm addressing my comment to you as I see you have been responding to former comments on this thread.

I have installed the necessary packages on my mac laptop and have also downloaded, built, etc requirements for flare portal. During flare portal setup, I was prompted in the command line to make a log-in for the portal, which I did (username, email, and password). I then tried to access the portal through both `http://127.0.0.1:8000/` and `localhost:3000`, providing the login info I had just generated. Both times I entered my login info and clicked the button to continue to the portal and the window did nothing. Please advise.

For reference, I am on a mac device and have downloaded Android Studio, which I intend to use to test FLARe. I am first setting up the portal, as instructed, before using/testing the app via Android Studio.

Thank you
