### 2021-07-27 #3270 @McGregor14
Hello @KirstensGitHub, sorry for the late reply.

I have nudged @NathanHorrigan by email, so hopefully he will get back shortly.

@oliviaguest

I think it would be worth me briefly explaining the current situation and roles of the authors for future reference, too.

FLARe is a King’s College London project that has been going for around 5 years, basically creating and using different versions of the app for research. I have been the scientific lead on the 3rd and 4th versions (which includes the one currently under review). KLP was the scientific lead prior to me (versions 1 and 2) and continues to provide significant input on the project. TCE and GB have led the project since its inception and KSY is an expert on the fear conditioning paradigm who has given us a lot of advice throughout the latest round of development.

Having previously worked with a different industry partner, for the most recent version of the app we went through a full rebuild with Torchbox and have made the software open source, hence our submission to JOSS. Nathan was the primary developer on the project, working on both the app and portal. MC initially developed the portal before Nathan took over both elements towards the end. BH was the project manager overseeing our day-to-day work with Torchbox. PV was in charge of strategy, helping ensure FLARe met the expectations of our team and our funders.

Nathan has since left Torchbox and moved to a new company that works on US time, as well as completing his undergraduate degree here in the UK. I think this has, understandably, made it difficult for him to respond promptly. As you can see from the roles outlined above, our team (at King’s College London) have a lot of expertise in using the product to conduct research and the science underlying it, but are not best placed to advise on the technical issues being addressed so far, here.

Hope this all makes sense and let me know if you have any questions.
