### 2021-08-11 #3270 @Morsey187
Hi @KirstensGitHub,

It looks like the authentication ("must_accept_terms" decorator) is set to compare against a field ("researcher_terms_updated_at") that doesn't have a value set by default, I presume this was to prevent people from signing up on the production site before any terms and conditions were set, however this should be handled automatically for local development or mentioned in the setup instructions at least.

I've written a quick fix below, but in the meantime I'll update the code, so we can avoid having to carry out this step for local builds.

Fix:

Within `fab sh` run `./manage.py shell`, then paste the code snipppet below and press enter. Then try accepting the terms and condtions again and you should be redirected to the dashboard.

```
from flare_portal.site_config.models import SiteConfiguration
from datetime import datetime
config = SiteConfiguration.get_solo()
config.researcher_terms_updated_at = datetime.now()

exit()
```

I've also had a look at the Portal and App repositories today and have made some notes on the documentation, so hopefully later this afternoon I can open some pull requests to address issues like "fab up" being the wrong command as well as add some further details to make setting up the FLARe app easier in relation to andriod and ios.

Thanks,
Ben
