+++ src/utils/exampleExperiment.ts
@@ -219,7 +219,7 @@ export const exampleExperimentData: Experiment = {
           didPayAttention: true,
           taskEnvironment: true,
           wasQuiet: true,
-          wasAlone: true,
+          wasNotAlone: true,
           wasInterrupted: true,
         },
       },
