### 2021-09-09 #3270 @EleonoraVagnoni
This is what I get if I run node -v
