### 2021-10-19 #3270 @Morsey187
Hi @EleonoraVagnoni,
It looks like you need to install the missing dependencies listed in that screenshot, for IOS that would be watchman which you can do by following the instructions [here](https://facebook.github.io/watchman/docs/install.html), or if you have [homebrew](https://brew.sh/) installed you can do this by running `brew install watchman`.
Similarly with the Andriod build, you'll have to install the listed dependencies that are crossed out.

Thanks,
Ben
