### 2021-11-26 #3270 @peircej
For my part, I'm stuck ticking the boxes for the following:
- **Example usage:** I could see any documentation showing usage.
- **Functionality documentation.** As above. For the phone app I couldn't see any signs of documentation other than how to build/install. For the portal there is a mention that, after installing the portal docs can be viewed at localhost but it would surely be worth providing documentation online. If the authors want users to get involved then being able to see what they're doing before committing to a substantial install process is important I think
- **Community guidelines.** No sign of these. The issues lists for the 2 projects are essentially unused. No suggestion of a forum/discussion site for users or contributors
- **State of the field.** I couldn't see any real discussion of the alternative options. There are many platforms out there with simpler processes for the user to run online behavioural studies which presumably include, but would not be limited to, fear conditioning. For instance, lab.js, jsPsych or PsychoPy can all generate a very wide range of behavioural studies and deploy them to online servers to run large numbers of participants at a time.

Overall I was left with a sense that this is a project built for the existing group that wrote and uses the software, and therefore doesn't require the documentation, rather than a one that expects a broader audience or community. To become an active package with wider appeal I would suggest working on publishing documentation about the software so that users can decide if it's the right package for them. Cross-linking the different sites will help that. Currently there's a webpage at https://www.kcl.ac.uk/research/flare but it doesn't point to the repositories or any documentation, and the readme files don't point back to that page.
