### 2021-11-27 #3270 @EleonoraVagnoni
Hello,
sorry for the late reply. I have tried to run brew install watchman. I am still stuck.
I think it is too difficult for a person with limited coding experience to download the software. I have been looking online for the errors I got but I have been copying and pasting lines without really understanding what I was doing. Even with your step-by-step guidance, I was not able to download it and use it. The documentation available is not enough for a person with limited coding experience.
