### 2021-11-30 #3270 @McGregor14
Yes, thanks all for the feedback. I think I was too focused on the author submission guidelines and should have spent time checking the reviewer checklist to see where the software stood and what was practically required to be classed as open-source. I will do my best to work with the team to fix the issues you have raised and will get back to you once I feel they have been addressed.

Thanks again for all your time so far.
