### 2022-02-03 #3270 @McGregor14
Hello @oliviaguest,

I have discussed the issues outlined above with my supervisor and, unfortunately, we have decided that it is not currently feasible for us to address them all in the foreseeable future. This is due to a number of reasons, mainly funding and capacity within our team (which is currently a bit stretched due to people moving on). I do apologise for the time this has taken you and the reviewers. I am a big fan of what the journal is doing and I appreciate how much work goes into any review, let alone that of potentially unfamiliar code/software. Hopefully, at some point, we can gain funding to work specifically on the software in order to make it more accessible to a broader audience, rather than just providing access to the codebase. Maybe then we can try another submission to JOSS.

Thank you and all the best,

Tom
