### 2022-02-03 #3270 @peircej
Sorry to see it came out that way (for now) but I think this is probably the right call. @McGregoer14 the work that's been done on the package is great and clearly very substantial but I think these requirements and expectations are there to get packages over that next line to becoming sustainable community projects. It's really just documentation that's missing here. Hopefully you'll be able to add the additional pieces and resubmit later.
